Responding to more comments

+++ dataset/src/utils/GraphInfo.h
@@ -30,7 +30,7 @@ class GraphInfo {
   void insertNode(uint64_t node_id, std::vector<float> features,
                   std::vector<uint64_t> neighbors);
 
-  uint32_t featureDim() const { return _feature_dim; }
+  uint64_t featureDim() const { return _feature_dim; }
 
  private:
   uint64_t _feature_dim;
